update code, reorg and cleanup for manipulation code
